+++ R/plot_wham_output.R
@@ -119,6 +119,8 @@ plot_wham_output <- function(mod, dir.main = getwd(), out.type = 'html', res = 7
     plot.M(mod)
     plot.tile.age.year(mod, type="selAA")
     plot.tile.age.year(mod, type="MAA")
+    plot_q_prior_post(mod) #flag inside to plot if prior is being used. 
+    plot_q(mod)
     if(!all(mod$env$data$Ecov_model == 0) & mod$is_sdrep) plot.ecov(mod)
     dev.off()
 
@@ -156,6 +158,10 @@ plot_wham_output <- function(mod, dir.main = getwd(), out.type = 'html', res = 7
 
   if(out.type %in% c('png','html')){
     dir.plots <- file.path(dir.main, "plots_png")
+    #This removes any plots from previous models. Some plots that are left in the directories may not pertain to the current model.
+    if(dir.exists(dir.plots)) {
+      file.remove(file.path(dir.plots, list.files(dir.plots, recursive = TRUE)))
+    }
     dir.create(dir.plots, showWarnings = FALSE)
 
     # PNG input_data -----------------
@@ -282,6 +288,8 @@ plot_wham_output <- function(mod, dir.main = getwd(), out.type = 'html', res = 7
     dev.off()
     plot.tile.age.year(mod, type="selAA", do.png=TRUE, fontfam=fontfam, od=dir.res)
     plot.tile.age.year(mod, type="MAA", do.png=TRUE, fontfam=fontfam, od=dir.res)
+    plot_q_prior_post(mod, do.png=TRUE, fontfam=fontfam, od=dir.res) #flag inside to plot if prior is being used. 
+    plot_q(mod, do.png=TRUE, fontfam=fontfam, od=dir.res)
     if(!all(mod$env$data$Ecov_model == 0) & mod$is_sdrep) plot.ecov(mod, do.png=TRUE, fontfam=fontfam, od=dir.res, res=res)
 
     # PNG reference points -----------------